Handled all of crud for workouts (only delete remains)

+++ src/Core/ValueObjects/Set/Weight.cs
@@ -22,7 +22,7 @@ public class Weight : ValueObject
         return weight.Value;
     }
 
-    public static implicit operator Weight(int value)
+    public static implicit operator Weight(double value)
     {
         return new Weight(value);
     }
